% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof.R
\name{cof}
\alias{cof}
\title{Consequences of Failure}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
cof(financial_cof, safety_cof, environmental_cof, network_cof)
}
\arguments{
\item{financial_cof}{Numeric. Financial consequences of failure.}

\item{safety_cof}{Numeric. Safety consequences of failure.}

\item{environmental_cof}{Numeric. Environmental consequences of failure.}

\item{network_cof}{Numeric. Network cost of failure.}
}
\value{
Numeric. Consequences of failure.
}
\description{
This function calculates consequences of failure
(cf.section 7, page 75, CNAIM, 2021).
}
