% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_discounted_pof.R
\name{cum_discounted_pof}
\alias{cum_discounted_pof}
\title{cumulative discounted PoF}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
cum_discounted_pof(pof, n, i)
}
\arguments{
\item{pof}{The probability of failure of the asset}

\item{n}{number of future years considered}

\item{i}{number of years subsequent to current year (where current year is year 0)}
}
\description{
This function calculates the cumulative discounted PoF.
See section 5.5 on page 32 in CNAIM (2021).
}
\examples{

}
