% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duty_factor_cables_u66kv.R
\name{duty_factor_cables_u66kv}
\alias{duty_factor_cables_u66kv}
\title{Duty Factor for all cables under 66kV}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/system/files/docs/2017/05/dno_common_network_asset_indices_methodology_v1.1.pdf}
}
\usage{
duty_factor_cables_u66kv(
  utilisation_pct = "Default",
  operating_voltage_pct = "Default",
  voltage_level = "EHV"
)
}
\arguments{
\item{utilisation_pct}{Numeric. The max percentage of utilisation
under normal operating conditions.}

\item{operating_voltage_pct}{Numeric. The percentage of operating/design
voltage.}

\item{voltage_level}{String. Specify the voltage level. Options:
\code{voltage_level = c("EHV", "LV & HV")}.
Choose \code{"EHV"} for 33-66kV cables and \code{"LV & HV"} for cables under
33kV.}
}
\value{
Numeric. Duty factor for cables under 66kV.
}
\description{
This function calculates the duty factor for under
cables under 66kV depending on the maximum
percentage utilisation under normal operating conditions.
The duty factor is used in the deriviation of the expected life of an asset.
See e.g. \code{\link{expected_life}}(). For more general information about
the derivation of the duty factor see section 6.6 on page 47 in CNAIM (2017)
}
\examples{
duty_factor_cables_u66kv(utilisation_pct = "Default",
operating_voltage_pct = "Default",
voltage_level = "EHV")
}
