% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duty_factor_transformer_11kv.R
\name{duty_factor_transformer_11kv}
\alias{duty_factor_transformer_11kv}
\title{Duty Factor for a 6.6/11 kV Transformer}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/system/files/docs/2017/05/dno_common_network_asset_indices_methodology_v1.1.pdf}
}
\usage{
duty_factor_transformer_11kv(utilisation_pct = "Default")
}
\arguments{
\item{utilisation_pct}{Numeric. The max percentage of utilisation
under normal operating conditions for a 6.6/11 kV transformer.}
}
\value{
Numeric. Duty factor for a 6.6/11 kV transformer.
}
\description{
This function calculates the duty factor for a 6.6/11 kV transformer
depending on the maximum percentage utilisation under normal operating
conditions.
The duty factor is used in the deriviation of the expected life of an asset.
See e.g. \code{\link{expected_life}}(). For more general information about
the derivation of the duty factor see section 6.6 on page 47 in CNAIM (2017)
}
\examples{
duty_factor_transformer_11kv(utilisation_pct = 95)
}
