% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_board_04kv.R
\name{environmental_cof_board_04kv}
\alias{environmental_cof_board_04kv}
\title{Environmental cost of Failure for 0.4kV Board}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
environmental_cof_board_04kv(gb_ref_given = NULL)
}
\arguments{
\item{gb_ref_given}{optional parameter to use custom reference values}
}
\description{
This function calculates environmental consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Environmental consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().#' @return Numeric. Financial consequences of failure for 0.4kV board
Outputted in DKK
}
\examples{
environmental_cof_board_04kv()
}
