% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ohl_cond_50kv.R
\name{environmental_cof_ohl_cond_50kv}
\alias{environmental_cof_ohl_cond_50kv}
\title{Environmental cost of Failure for 50kV Overhead Line Conductors}
\usage{
environmental_cof_ohl_cond_50kv()
}
\description{
This function calculates environmental consequences of failure
Outputted in DKK
}
\examples{
environmental_cof_ohl_cond_50kv()
}
