% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ohl_fittings_50kv.R
\name{environmental_cof_ohl_fittings_50kv}
\alias{environmental_cof_ohl_fittings_50kv}
\title{Environmental cost of Failure for 50kV Fittings}
\usage{
environmental_cof_ohl_fittings_50kv()
}
\value{
Numeric. Financial consequences of failure for 50kv fittings
Outputted in DKK.
}
\description{
This function calculates environmental consequences of failure
Environmental consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
environmental_cof_ohl_fittings_50kv()
}
