% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_poles_ohl_support_50kv.R
\name{environmental_cof_poles_ohl_support_50kv}
\alias{environmental_cof_poles_ohl_support_50kv}
\title{Environmental cost of Failure for Poles OHL Support 50kV}
\usage{
environmental_cof_poles_ohl_support_50kv()
}
\description{
This function calculates environmental consequences of failure
Environmental consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().#' @return Numeric.
Financial consequences of failure for Poles OHL support 50kV
Outputted in DKK.
}
\examples{
environmental_cof_poles_ohl_support_50kv()
}
