% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_sub_cables.R
\name{environmental_cof_sub_cables}
\alias{environmental_cof_sub_cables}
\title{Environmental cost of Failure for sub cables}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2017:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
environmental_cof_sub_cables(sub_cable_asset_category)
}
\arguments{
\item{sub_cable_asset_category}{String The type of Submarine cable asset category
Options: \code{sub_cable_asset_category = c("HV Sub Cable",
"EHV Sub Cable", "132kV Sub Cable")}.}
}
\value{
Numeric. Financial consequences of failure for sub cables
}
\description{
This function calculates environmental consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Environmental consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
environmental_cof_sub_cables(sub_cable_asset_category = "HV Sub Cable")
}
