% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_tower_ohl_support_50kv.R
\name{environmental_cof_tower_ohl_support_50kv}
\alias{environmental_cof_tower_ohl_support_50kv}
\title{Environmental cost of Failure for Tower OHL Support 50 kV}
\usage{
environmental_cof_tower_ohl_support_50kv()
}
\value{
Numeric. Financial consequences of failure for tower ohl support 50 kV
Outputted in DKK.
}
\description{
This function calculates environmental consequences of failure
Environmental consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
environmental_cof_tower_ohl_support_50kv()
}
