% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_towers.R
\name{environmental_cof_towers}
\alias{environmental_cof_towers}
\title{Environmental cost of Failure for towers}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
environmental_cof_towers(tower_asset_category)
}
\arguments{
\item{tower_asset_category}{String The type of tower asset category
Options: \code{tower_asset_category = c("33kV Tower", "66kV Tower",
"132kV Tower")}.}
}
\value{
Numeric. Financial consequences of failure for towers
}
\description{
This function calculates environmental consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Environmental consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
environmental_cof_towers(tower_asset_category = "33kV Tower")
}
