% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_cables_04_10kv.R
\name{financial_cof_cables_04_10kv}
\alias{financial_cof_cables_04_10kv}
\title{Financial cost of Failure for 0.4kV and 10kV UG Cables}
\usage{
financial_cof_cables_04_10kv(hv_asset_category)
}
\value{
Numeric. Financial consequences of failure for 0.4kV and 10kV UG cables
}
\description{
This function calculates financial consequences of failure
Outputted in DKK
\code{hv_asset_category = c("10kV UG Cable (Oil)",
"10kV UG Cable (Non Pressurised)", "0.4kV UG Cable (Non Pressurised)"}.
}
\examples{
financial_cof_cables_04_10kv(hv_asset_category = "10kV UG Cable (Oil)")
}
