% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ehv_fittings.R
\name{financial_cof_ehv_fittings}
\alias{financial_cof_ehv_fittings}
\title{Financial cost of Failure for EHV/132kV fittings}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
financial_cof_ehv_fittings(
  ehv_asset_category,
  type_financial_factor_criteria,
  access_factor_criteria
)
}
\arguments{
\item{ehv_asset_category}{String The type of EHV Fittings asset category}

\item{type_financial_factor_criteria}{String. Type Financial factor criteria for EHV fittings}

\item{access_factor_criteria}{String. Asses Financial factor criteria for EHV fittings
setting (cf. table 214, page 164, CNAIM, 2017).}
}
\value{
Numeric. Financial consequences of failure for EHV fittings
}
\description{
This function calculates financial consequences of failure
(cf. section 7.3, page 75, CNAIM, 2017). Financial consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
financial_cof_ehv_fittings(ehv_asset_category = "33kV Fittings",
type_financial_factor_criteria = "Tension",
access_factor_criteria = "Type A")
}
