% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ohl_cond_50kv.R
\name{financial_cof_ohl_cond_50kv}
\alias{financial_cof_ohl_cond_50kv}
\title{Financial cost of Failure for 50kV Overhead Line Conductors}
\usage{
financial_cof_ohl_cond_50kv(access_factor_criteria)
}
\arguments{
\item{access_factor_criteria}{String. Asses Financial factor criteria for
Overhead Line Conductors.
Options: \code{access_factor_criteria = c("Type A", "Type B")}.}
}
\value{
Numeric. Financial consequences of failure for Overhead Line Conductors
}
\description{
This function calculates financial consequences of failure
Outputted in DKK
}
\examples{
financial_cof_ohl_cond_50kv(
access_factor_criteria = "Type A")
}
