% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ohl_fittings_50kv.R
\name{financial_cof_ohl_fittings_50kv}
\alias{financial_cof_ohl_fittings_50kv}
\title{Financial cost of Failure for 50kV Fittings}
\usage{
financial_cof_ohl_fittings_50kv(
  type_financial_factor_criteria,
  access_factor_criteria
)
}
\arguments{
\item{type_financial_factor_criteria}{String. Type Financial factor criteria for EHV fittings
Options: \code{type_financial_factor_criteria = c("Suspension", "Tension")}.}

\item{access_factor_criteria}{String. Asses Financial factor criteria for EHV fittings
setting. Options: \code{access_factor_criteria = c("Type A", "Type B")}.}
}
\value{
Numeric. Financial consequences of failure for EHV fittings
}
\description{
This function calculates financial consequences of failure.
Financial consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK
}
\examples{
financial_cof_ohl_fittings_50kv(
type_financial_factor_criteria = "Tension",
access_factor_criteria = "Type A")
}
