% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_relay.R
\name{financial_cof_relay}
\alias{financial_cof_relay}
\title{Financial cost of Failure for Relays}
\usage{
financial_cof_relay(access_factor_criteria)
}
\arguments{
\item{access_factor_criteria}{String. Asses Financial factor criteria for relay
setting. Options: \code{access_factor_criteria = c("Type A", "Type B", "Type C")}.}
}
\description{
This function calculates financial consequences of failure
Financial consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
\examples{
financial_cof_relay(access_factor_criteria = "Type A")
}
