% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_serviceline.R
\name{financial_cof_serviceline}
\alias{financial_cof_serviceline}
\title{Financial cost of Failure for Service Lines}
\usage{
financial_cof_serviceline(gb_ref_given = NULL)
}
\arguments{
\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Financial consequences of failure for service line
}
\description{
This function calculates financial consequences of failure
Outputted in DKK
}
\examples{
financial_cof_serviceline()
}
