% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_submarine_cables_10kv.R
\name{financial_cof_submarine_cables_10kv}
\alias{financial_cof_submarine_cables_10kv}
\title{Financial cost of Failure for 10kV Submarine Cables}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
financial_cof_submarine_cables_10kv()
}
\value{
Numeric. Financial consequences of failure for 10kV submarine cables
}
\description{
This function calculates financial consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Financial consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
\examples{
financial_cof_submarine_cables_10kv()
}
