% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_switchgear_30_60kv.R
\name{financial_cof_switchgear_30_60kv}
\alias{financial_cof_switchgear_30_60kv}
\title{Financial cost of Failure for 30kV and 60kV Switchgear}
\usage{
financial_cof_switchgear_30_60kv(ehv_asset_category, access_factor_criteria)
}
\arguments{
\item{ehv_asset_category}{String The type of 30kV and 60kV switchgear
Options: \code{ehv_asset_category = c("30kV", "60kV")}.}

\item{access_factor_criteria}{String. Asses Financial factor criteria for 30kV and 60kV switchgear
setting. Options: \code{access_factor_criteria = c("Type A", "Type B", "Type C")}.}
}
\value{
Numeric. Financial consequences of failure for 30kV and 60kV switchgear
}
\description{
This function calculates financial consequences of failure
Financial consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
