% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_tower_ohl_support_50kv.R
\name{financial_cof_tower_ohl_support_50kv}
\alias{financial_cof_tower_ohl_support_50kv}
\title{Financial cost of Failure for Tower OHL Support 50 kV}
\usage{
financial_cof_tower_ohl_support_50kv(
  type_financial_factor_criteria,
  access_factor_criteria,
  gb_ref_given = NULL
)
}
\arguments{
\item{type_financial_factor_criteria}{String The type financial factor for Tower
Options: \code{type_financial_factor_criteria = c("Suspension",
"Tension", "Terminal")}.}

\item{access_factor_criteria}{String. Asses Financial factor criteria for Tower
Options: \code{access_factor_criteria = c("Type A", "Type B")}.}

\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Financial consequences of failure for tower ohl support 50 kV
}
\description{
This function calculates financial consequences of failure
Financial consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
\examples{
financial_cof_tower_ohl_support_50kv(
type_financial_factor_criteria = "Suspension",
access_factor_criteria = "Type A")
}
