% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{matrix_adjusted_circles}
\alias{matrix_adjusted_circles}
\title{Adjust circles for matrix visualization}
\usage{
matrix_adjusted_circles(risk_data_matrix, dots_vector, dot_radius)
}
\arguments{
\item{risk_data_matrix}{Long format matrix data.}

\item{dots_vector}{Coordinates of the dots.}

\item{dot_radius}{Radius of the dots.}
}
\description{
This function manipulates the data structure before inputting
into javascript D3 risk matrix visualization
}
