% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{matrix_adjusted_intervals}
\alias{matrix_adjusted_intervals}
\title{Adjust banding for matrix visualization}
\usage{
matrix_adjusted_intervals(risk_data_matrix, x_intervals, y_intervals)
}
\arguments{
\item{risk_data_matrix}{Long format matrix data.}

\item{x_intervals}{An array of x spacing in percent (sum to 100)}

\item{y_intervals}{An array of y spacing in percent (sum to 100)}
}
\description{
This function manipulates the data structure before inputting
into javascript D3 risk matrix visualization
}
