% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_cables_60_30kv.R
\name{network_cof_cables_60_30kv}
\alias{network_cof_cables_60_30kv}
\title{Network cost of Failure for 30-60 kV UG cables}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
network_cof_cables_60_30kv(
  ehv_asset_category,
  actual_load_mva,
  secure = T,
  gb_ref_given = NULL
)
}
\arguments{
\item{ehv_asset_category}{Asset category for the analysis}

\item{actual_load_mva}{Numeric. The actual load on the asset}

\item{secure}{Boolean If the asset is in a secure network or not}

\item{gb_ref_given}{optional parameter to use custo,e reference values}
}
\value{
Numeric. Network cost of failure.
}
\description{
This function calculates network cost of failure for
all asset categories exclusive the assets EHV and 132kV transformers.
(cf. section 7.6, page 87, CNAIM, 2021). Network cost of failure
is used in the derivation of consequences of failure see \code{\link{cof}}().
\code{ehv_asset_category = c("30kV UG Cable (Gas)", "60kV UG Cable (Gas)",
"30kV UG Cable (Non Pressurised)", "60kV UG Cable (Non Pressurised)",
"30kV UG Cable (Oil)", "60kV UG Cable (Oil)")
}. The default setting is
\code{ehv_asset_category = "60kV UG Cable (Gas)"}.
}
\examples{
network_cof_cables_60_30kv(ehv_asset_category = "30kV UG Cable (Oil)",
actual_load_mva = 15)
}
