% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ohl_fittings_50kv.R
\name{network_cof_ohl_fittings_50kv}
\alias{network_cof_ohl_fittings_50kv}
\title{Network cost of Failure for 50kV Fittings}
\usage{
network_cof_ohl_fittings_50kv(actual_load_mva, secure = T)
}
\arguments{
\item{actual_load_mva}{Numeric. The actual load on the asset}

\item{secure}{Boolean If the asset is in a secure network or not}
}
\value{
Numeric. Network cost of failure.
}
\description{
This function calculates network cost of failure for
50kV fittings
Network cost of failure is used in the derivation of consequences of failure.
\code{\link{cof}}().
Outputted in DKK.
}
\examples{
network_cof_ohl_fittings_50kv(
actual_load_mva = 15)
}
