% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_serviceline.R
\name{network_cof_serviceline}
\alias{network_cof_serviceline}
\title{Network cost of Failure for Service Lines}
\usage{
network_cof_serviceline(actual_load_mva, secure = T, gb_ref_given = NULL)
}
\arguments{
\item{actual_load_mva}{Numeric. The actual load on the asset}

\item{secure}{Boolean If the asset is in a secure network or not}

\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Network cost of failure.
}
\description{
This function calculates network cost of failure for
service line, outputted in DKK
}
\examples{
network_cof_serviceline(actual_load_mva = 0.5)
}
