% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_switchgear_30_60kv.R
\name{network_cof_switchgear_30_60kv}
\alias{network_cof_switchgear_30_60kv}
\title{Network cost of Failure for 30kV and 60kV Switchgear}
\usage{
network_cof_switchgear_30_60kv(
  ehv_asset_category,
  actual_load_mva,
  secure = T,
  gb_ref_given = NULL
)
}
\arguments{
\item{ehv_asset_category}{String The type of 30kV and 60kV switchgear category
Options: \code{ehv_asset_category = c("30kV", "60kV")}.}

\item{actual_load_mva}{Numeric. The actual load on the asset}

\item{secure}{Boolean If the asset is in a secure network or not}

\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Network cost of failure.
}
\description{
This function calculates network cost of failure for
30kV and 60kV switchgear.
Network cost of failureis used in the derivation
 of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
\examples{
network_cof_switchgear_30_60kv(ehv_asset_category = "30kV",
actual_load_mva = 15)
}
