% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_cables_10kv_apb.R
\name{pof_cables_10kv_apb}
\alias{pof_cables_10kv_apb}
\title{Current Probability of Failure for Danish 10kV APB cables}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
pof_cables_10kv_apb(
  utilisation_pct = "Default",
  operating_voltage_pct = "Default",
  sheath_test = "Default",
  partial_discharge = "Default",
  fault_hist = "Default",
  reliability_factor = "Default",
  age
)
}
\arguments{
\item{utilisation_pct}{Numeric. The max percentage of utilisation
under normal operating conditions.}

\item{operating_voltage_pct}{Numeric. The ratio in percent of
operating/design voltage.}

\item{sheath_test}{String. Only applied for non pressurised cables.
Indicating the state of the sheath. Options:
\code{sheath_test = c("Pass", "Failed Minor", "Failed Major",
"Default")}. See page 153, table 168 in CNAIM (2021).}

\item{partial_discharge}{String. Only applied for non pressurised cables.
Indicating the level of partial discharge. Options:
\code{partial_discharge = c("Low", "Medium", "High",
 "Default")}. See page 153, table 169 in CNAIM (2021).}

\item{fault_hist}{Numeric. Only applied for non pressurised cables.
The calculated fault rate for the cable in the period per kilometer.
A setting of \code{"No historic faults recorded"}
indicates no fault. See page 153, table 170 in CNAIM (2021).}

\item{reliability_factor}{Numeric. \code{reliability_factor}
shall have a value between 0.6 and 1.5. A setting of \code{"Default"}
 sets the \code{reliability_factor} to 1.
See section 6.14 on page 73 in CNAIM (2021).}

\item{age}{Numeric. The current age in years of the cable.}
}
\value{
Numeric. Current probability of failure
per annum for 10 kV apb cables.
}
\description{
This function calculates the current
annual probability of failure per kilometer for a Danish 10 kV APB cable.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function. For more information about the
probability of failure function see section 6
on page 34 in CNAIM (2021).
}
\examples{
# Current annual probability of failure for 10-20kV cable, APB, 50 years old
pof_cables_10kv_apb_result <-
pof_cables_10kv_apb(
utilisation_pct = 80,
operating_voltage_pct = 60,
sheath_test = "Default",
partial_discharge = "Default",
fault_hist = "Default",
reliability_factor = "Default",
age = 50) * 100

paste0(sprintf("Probability of failure \%.4f", pof_cables_10kv_apb_result),
" percent per annum")
}
