% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_cables_04kv_pex.R
\name{pof_future_cables_04kv_pex}
\alias{pof_future_cables_04kv_pex}
\title{Future Probability of Failure for 0.4kV UG PEX Non Pressurised Cables}
\usage{
pof_future_cables_04kv_pex(
  utilisation_pct = "Default",
  operating_voltage_pct = "Default",
  sheath_test = "Default",
  partial_discharge = "Default",
  fault_hist = "Default",
  reliability_factor = "Default",
  age,
  k_value = 0.0658,
  c_value = 1.087,
  normal_expected_life = 80,
  simulation_end_year = 100
)
}
\arguments{
\item{utilisation_pct}{Numeric. The max percentage of utilisation
under normal operating conditions.}

\item{operating_voltage_pct}{Numeric. The ratio in percent of
operating/design voltage.}

\item{sheath_test}{String. Only applied for non pressurised cables.
Indicating the state of the sheath. Options:
\code{sheath_test = c("Pass", "Failed Minor", "Failed Major",
"Default")}.}

\item{partial_discharge}{String. Only applied for non pressurised cables.
Indicating the level of partial discharge. Options:
\code{partial_discharge = c("Low", "Medium", "High",
 "Default")}.}

\item{fault_hist}{Numeric. Only applied for non pressurised cables.
The calculated fault rate for the cable in the period per kilometer.
A setting of \code{"No historic faults recorded"}
indicates no fault.}

\item{reliability_factor}{Numeric. \code{reliability_factor}
shall have a value between 0.6 and 1.5. A setting of \code{"Default"}
 sets the \code{reliability_factor} to 1.
See section 6.14 on page 73 in CNAIM (2021).}

\item{age}{Numeric. The current age in years of the cable.}

\item{k_value}{Numeric. \code{k_value = 0.0658} by default.}

\item{c_value}{Numeric. \code{c_value = 1.087} by default.
The default value is accordingly to the CNAIM standard see page 110}

\item{normal_expected_life}{Numeric. \code{normal_expected_life = 80} by default.}

\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
DataFrame. Future probability of failure
along with future health score
}
\description{
This function calculates the future
annual probability of failure per kilometer for a 0.4kV PEX non Pressurised cables
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# future annual probability of failure for 0.4kV cable pex, 50 years old
pof_future_cables_04kv_pex(
utilisation_pct = 80,
operating_voltage_pct = 60,
sheath_test = "Default",
partial_discharge = "Default",
fault_hist = "Default",
reliability_factor = "Default",
age = 50,
k_value = 0.0658,
c_value = 1.087,
normal_expected_life = 80,
simulation_end_year = 100)
}
