% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_ohl_cond_50kv.R
\name{pof_future_ohl_cond_50kv}
\alias{pof_future_ohl_cond_50kv}
\title{Future Probability of Failure for 50kV OHL Conductors}
\usage{
pof_future_ohl_cond_50kv(
  sub_division = "Cu",
  placement = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  conductor_samp = "Default",
  corr_mon_survey = "Default",
  visual_cond = "Default",
  midspan_joints = "Default",
  reliability_factor = "Default",
  k_value = 0.008,
  c_value = 1.087,
  normal_expected_life = "Default",
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric. Future probability of failure
per annum per kilometer.
}
\description{
This function calculates the future
annual probability of failure per kilometer 50kV OHL conductors.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# Future annual probability of failure for 50kV OHL (Tower Line) Conductor
}
