% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_ohl_fittings_50kv.R
\name{pof_future_ohl_fittings_50kv}
\alias{pof_future_ohl_fittings_50kv}
\title{Future Probability of Failure for 50 kV Fittings}
\usage{
pof_future_ohl_fittings_50kv(
  placement = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  measured_condition_inputs,
  observed_condition_inputs,
  reliability_factor = "Default",
  k_value = 0.0096,
  c_value = 1.087,
  normal_expected_life = 40,
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.code}
}
\value{
Numeric array. Future probability of failure
per annum per kilometre for 50 kV fittings.
}
\description{
This function calculates the future
annual probability of failure per kilometer for a 50 kV fittings.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# Future annual probability of failure for 50kV fittings
}
