% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_poles_ohl_support_50kv.R
\name{pof_future_poles_ohl_support_50kv}
\alias{pof_future_poles_ohl_support_50kv}
\title{Future Probability of Failure for Poles OHL support 50 kV}
\usage{
pof_future_poles_ohl_support_50kv(
  sub_division = "Wood",
  placement = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  measured_condition_inputs,
  observed_condition_inputs,
  reliability_factor = "Default",
  k_value = 0.0285,
  c_value = 1.087,
  normal_expected_life = "Default",
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric array. Future probability of failure
per annum per kilometre for Poles OHL support 50 kV.
}
\description{
This function calculates the future
annual probability of failure per kilometer for a Poles OHL support 50 kV.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# Future annual probability of failure for Poles OHL support 50 kV
}
