% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_rtu.R
\name{pof_future_rtu}
\alias{pof_future_rtu}
\title{Future Probability of Failure for RTU}
\usage{
pof_future_rtu(
  placement = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  measured_condition_inputs,
  observed_condition_inputs,
  reliability_factor = "Default",
  k_value = 0.128,
  c_value = 1.087,
  normal_expected_life = 20,
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric. Current probability of failure
per annum
}
\description{
This function calculates the future
annual probability of failure RTU.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
 # future annual probability of failure for RTU
}
