% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_serviceline.R
\name{pof_future_serviceline}
\alias{pof_future_serviceline}
\title{Future Probability of Failure for Service Line}
\usage{
pof_future_serviceline(
  utilisation_pct = "Default",
  operating_voltage_pct = "Default",
  sheath_test = "Default",
  partial_discharge = "Default",
  fault_hist = "Default",
  reliability_factor = "Default",
  age,
  k_value = 0.0329,
  c_value = 1.087,
  normal_expected_life = 75,
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric array. Future probability of failure
per annum for service line.
}
\description{
This function calculates the future
annual probability of failure per kilometer for a service line
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# future annual probability of failure for service line, 50 years old
}
