% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_submarine_cables_10kv_oil.R
\name{pof_future_submarine_cables_10kv_oil}
\alias{pof_future_submarine_cables_10kv_oil}
\title{Future Probability of Failure for 10kV Oil Submarine Cables}
\usage{
pof_future_submarine_cables_10kv_oil(
  utilisation_pct = "Default",
  operating_voltage_pct = "Default",
  topography = "Default",
  sitution = "Default",
  wind_wave = "Default",
  intensity = "Default",
  landlocked = "no",
  sheath_test = "Default",
  partial_discharge = "Default",
  fault_hist = "Default",
  condition_armour = "Default",
  age,
  reliability_factor = "Default",
  k_value = 2.0944,
  c_value = 1.087,
  normal_expected_life = 60,
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric array. Future probability of failure
per annum per kilometre for 10kV submarine cables
}
\description{
This function calculates the future
annual probability of failure per kilometer for a 10kV Oil submarine cables
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# Future annual probability of failure for 1 km 10kV Oil Sub Cable
pof_future_submarine_cables_10kv_oil(
utilisation_pct = "Default",
operating_voltage_pct = "Default",
topography = "Default",
sitution = "Default",
wind_wave = "Default",
intensity = "Default",
landlocked = "no",
sheath_test = "Default",
partial_discharge = "Default",
fault_hist = "Default",
condition_armour = "Default",
age = 10,
reliability_factor = "Default",
k_value = 0.0202,
c_value = 1.087,
normal_expected_life = 60,
simulation_end_year = 100)
}
