% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_submarine_cables_30_60kv_pex.R
\name{pof_future_submarine_cables_30_60kv_pex}
\alias{pof_future_submarine_cables_30_60kv_pex}
\title{Future Probability of Failure for 30kV and 60kV Non Pressurised Submarine Cables}
\usage{
pof_future_submarine_cables_30_60kv_pex(
  utilisation_pct = "Default",
  operating_voltage_pct = "Default",
  topography = "Default",
  sitution = "Default",
  wind_wave = "Default",
  intensity = "Default",
  landlocked = "no",
  sheath_test = "Default",
  partial_discharge = "Default",
  fault_hist = "Default",
  condition_armour = "Default",
  age,
  reliability_factor = "Default",
  k_value = 0.0202,
  c_value = 1.087,
  normal_expected_life = 60,
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric array. Future probability of failure
per annum per kilometre for 30kV and 60kV submarine cables
}
\description{
This function calculates the future
annual probability of failure per kilometer for a 30kV and 60kV Non Pressurised submarine cables
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# Future annual probability of failure for 1 km 30kV and 60kV non pressurised Sub Cable
}
