% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_switchgear_secondary_10kv.R
\name{pof_future_switchgear_secondary_10kv}
\alias{pof_future_switchgear_secondary_10kv}
\title{Future Probability of Failure for 10kV Switchgear Secondary}
\usage{
pof_future_switchgear_secondary_10kv(
  placement = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  measured_condition_inputs,
  observed_condition_inputs,
  reliability_factor = "Default",
  k_value = 0.0067,
  c_value = 1.087,
  normal_expected_life = 55,
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric. Current probability of failure per annum.
}
\description{
This function calculates the future
annual probability of failure 10kV switchgear secondary.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
 # future annual probability of failure for 10kV Switchgear secondary
}
