% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_tower_ohl_support_50kv.R
\name{pof_tower_ohl_support_50kv}
\alias{pof_tower_ohl_support_50kv}
\title{Current Probability of Failure for Towers OHL support 50kV}
\usage{
pof_tower_ohl_support_50kv(
  foundation_type = "Foundation - Fully Encased Concrete",
  paint_type = "Paint System - Paint",
  placement = "Default",
  number_of_operations = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  observed_condition_inputs_steelwork,
  observed_condition_inputs_paint,
  observed_condition_inputs_foundation,
  reliability_factor = "Default",
  k_value = 0.0545,
  c_value = 1.087,
  normal_expected_life = "Default"
)
}
\arguments{
\item{foundation_type}{String. Foundation type of the tower
\code{foundation_type = c("Foundation - Fully Encased Concrete",
 "Foundation - Earth Grillage")}}

\item{paint_type}{String. Paint type of the tower
\code{foundation_type = c(Paint System - Galvanising, Paint System - Paint )}}

\item{placement}{String. Specify if the asset is located outdoor or indoor.}

\item{number_of_operations}{Numeric Number of operations for the tower}

\item{altitude_m}{Numeric. Specify the altitude location for
the asset measured in meters from sea level.\code{altitude_m}
is used to derive the altitude factor. A setting of \code{"Default"}
will set the altitude factor to 1 independent of \code{asset_type}.}

\item{distance_from_coast_km}{Numeric. Specify the distance from the
coast measured in kilometers. \code{distance_from_coast_km} is used
to derive the distance from coast factor.
A setting of \code{"Default"} will set the
 distance from coast factor to 1 independent of \code{asset_type}.}

\item{corrosion_category_index}{Integer.
Specify the corrosion index category, 1-5.}

\item{age}{Numeric. The current age in years of the conductor.}

\item{observed_condition_inputs_steelwork}{Named list observed_conditions_input}

\item{observed_condition_inputs_paint}{Named list observed_conditions_input}

\item{observed_condition_inputs_foundation}{Named list observed_conditions_input
\code{conductor_samp = c("Low","Medium/Normal","High","Default")}.}

\item{reliability_factor}{Numeric. \code{reliability_factor}
shall have a value between 0.6 and 1.5. A setting of \code{"Default"}
 sets the \code{reliability_factor} to 1.
See section 6.14 on page 73 in CNAIM (2021).}

\item{k_value}{Numeric. \code{k_value = 0.0545} by default. This number is
given in a percentage. The default value is accordingly to the CNAIM standard
on p. 110.}

\item{c_value}{Numeric. \code{c_value = 1.087} by default.
The default value is accordingly to the CNAIM standard see page 110}

\item{normal_expected_life}{Numeric. \code{normal_expected_life = "Default"} by default.
The default value is accordingly to the CNAIM standard on page 107.}
}
\value{
Numeric. Current probability of failure
per annum per kilometer.
}
\description{
This function calculates the current
annual probability of failure per kilometer EHV for Towers OHL support 50kV
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# Current annual probability of failure for Towers
}
