% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_towers.R
\name{pof_towers}
\alias{pof_towers}
\title{Current Probability of Failure for Towers}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
pof_towers(
  tower_asset_category = "33kV Tower",
  foundation_type = "Foundation - Fully Encased Concrete",
  paint_type = "Paint System - Paint",
  placement = "Default",
  number_of_operations = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  observed_condition_inputs_steelwork,
  observed_condition_inputs_paint,
  observed_condition_inputs_foundation,
  reliability_factor = "Default"
)
}
\arguments{
\item{tower_asset_category}{String The type of Tower asset category}

\item{foundation_type}{String Foundation type of the tower}

\item{paint_type}{String Paint type of the tower}

\item{placement}{String. Specify if the asset is located outdoor or indoor.}

\item{number_of_operations}{Numeric Number of operations for the tower}

\item{altitude_m}{Numeric. Specify the altitude location for
the asset measured in meters from sea level.\code{altitude_m}
is used to derive the altitude factor. See page 107,
table 23 in CNAIM (2017). A setting of \code{"Default"}
will set the altitude factor to 1 independent of \code{asset_type}.}

\item{distance_from_coast_km}{Numeric. Specify the distance from the
coast measured in kilometers. \code{distance_from_coast_km} is used
to derive the distance from coast factor See page 106,
table 22 in CNAIM (2017). A setting of \code{"Default"} will set the
 distance from coast factor to 1 independent of \code{asset_type}.}

\item{corrosion_category_index}{Integer.
Specify the corrosion index category, 1-5.}

\item{age}{Numeric. The current age in years of the conductor.}

\item{observed_condition_inputs_steelwork}{Named list observed_conditions_input}

\item{observed_condition_inputs_paint}{Named list observed_conditions_input}

\item{observed_condition_inputs_foundation}{Named list observed_conditions_input
\code{conductor_samp = c("Low","Medium/Normal","High","Default")}.
See page 146-147, table 192 and 194 in CNAIM (2017).}

\item{reliability_factor}{Numeric. \code{reliability_factor}
shall have a value between 0.6 and 1.5. A setting of \code{"Default"}
 sets the \code{reliability_factor} to 1.
See section 6.14 on page 69 in CNAIM (2017).}
}
\value{
Numeric. Current probability of failure
per annum per kilometer.
}
\description{
This function calculates the current
annual probability of failure per kilometer EHV Switchgear
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function. For more information about the
probability of failure function see section 6
on page 30 in CNAIM (2017).
}
\examples{
# Current annual probability of failure for Towers
pof_towers(
tower_asset_category = "33kV Tower",
number_of_operations = "Default",
placement = "Default",
altitude_m = "Default",
distance_from_coast_km = "Default",
corrosion_category_index = "Default",
age = 10,
paint_type = "Paint System - Galvanising",
foundation_type = "Foundation - Earth Grillage",
observed_condition_inputs_steelwork =
list("tower_legs" = list("Condition Criteria: Observed Condition" = "Default"),
"tower_bracings" = list("Condition Criteria: Observed Condition" = "Default"),
"tower_crossarms" = list("Condition Criteria: Observed Condition" = "Default"),
"tower_peak" = list("Condition Criteria: Observed Condition" = "Default")),
observed_condition_inputs_paint =
list("paintwork_cond" = list("Condition Criteria: Observed Condition" = "Default")),
observed_condition_inputs_foundation =
list("foundation_cond" = list("Condition Criteria: Observed Condition" = "Default")),
reliability_factor = "Default")
}
