% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_calculation.R
\name{risk_calculation}
\alias{risk_calculation}
\title{Calculates risk and converts to matrix coordinates}
\usage{
risk_calculation(matrix_dimensions, id, pof, cof, asset_type)
}
\arguments{
\item{matrix_dimensions}{A data frame with the dimensions of the desired risk
matrix}

\item{id}{A string that describes the asset}

\item{pof}{The probability of failure of the asset}

\item{cof}{The consequences of failure of the asset}

\item{asset_type}{The asset type to be calculated for
class}
}
\description{
This function calculates monetary risk, given probability of
failure and consequence of failure inputs, as well as the desired risk matrix
dimensions.
}
