% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_matrix_points_plot.R
\name{risk_matrix_points_plot}
\alias{risk_matrix_points_plot}
\title{Make a risk matrix with individual asset points}
\usage{
risk_matrix_points_plot(risk_data_matrix, dots_vector, dot_radius)
}
\arguments{
\item{risk_data_matrix}{Long format matrix data.}

\item{dots_vector}{Coordinates of the dots.}

\item{dot_radius}{Radius of the dots.}
}
\description{
This function makes a D3 visualization of monetary risk with
each asset as a point on the grid.
}
