% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_matrix_structure.R
\name{risk_matrix_structure}
\alias{risk_matrix_structure}
\title{Makes a default risk matrix structure}
\usage{
risk_matrix_structure(cols, rows, value = NA)
}
\arguments{
\item{cols}{Number of columns}

\item{rows}{Number of rows}

\item{value}{Default value of each cell}
}
\description{
This function makes a simple matrix structure that can be used
as an input to the risk_matrix_points and risk_matrix_summary functions
}
