% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_matrix_summary_plot.R
\name{risk_matrix_summary_plot}
\alias{risk_matrix_summary_plot}
\title{Make a risk matrix with non-linear spacing}
\usage{
risk_matrix_summary_plot(
  risk_data_matrix,
  x_intervals = rep(20, 5),
  y_intervals = rep(25, 4)
)
}
\arguments{
\item{risk_data_matrix}{Long format matrix data.}

\item{x_intervals}{An array of x spacing in percent (sum to 100)}

\item{y_intervals}{An array of y spacing in percent (sum to 100)}
}
\description{
This function makes a D3 visualization of monetary risk with
non-linear x and y intervals.
}
