% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_cables_04_10kv.R
\name{safety_cof_cables_04_10kv}
\alias{safety_cof_cables_04_10kv}
\title{Safety cost of Failure for 0.4kV and 10kV UG Cables}
\usage{
safety_cof_cables_04_10kv(hv_asset_category)
}
\value{
Numeric. Financial consequences of failure for 0.4kV and 10kV UG cables
}
\description{
This function calculates safety consequences of failure
Outputted in DKK
\code{hv_asset_category = c("10kV UG Cable (Oil)",
"10kV UG Cable (Non Pressurised)", "0.4kV UG Cable (Non Pressurised)"}.
}
\examples{
safety_cof_cables_04_10kv(hv_asset_category = "10kV UG Cable (Oil)")
}
