% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ehv_fittings.R
\name{safety_cof_ehv_fittings}
\alias{safety_cof_ehv_fittings}
\title{Safety cost of Failure for EHV/132kV Fittings}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
safety_cof_ehv_fittings(
  ehv_asset_category,
  location_risk,
  type_risk,
  gb_ref_given = NULL
)
}
\arguments{
\item{ehv_asset_category}{String The type of EHV asset category
Options: \code{ehv_asset_category = c("33kV Fittings", "66kV Fittings",
 "132kV Fittings")}}

\item{location_risk}{String Type Financial factor criteria for EHV fittings
(cf. section D1.2.1, page 178, CNAIM, 2021).
\code{location_risk = c("Low", "Medium", "High")}.
The default setting is
\code{location_risk = "Medium"}.}

\item{type_risk}{String. Asses Financial factor criteria for EHV fittings
setting (cf. table 221, page 180, CNAIM, 2021).
\code{type_risk = c("Low", "Medium", "High")}.
The default setting is
\code{type_risk = "Medium"}.}

\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Financial consequences of failure for EHV fittings
}
\description{
This function calculates safety consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Safety consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
safety_cof_ehv_fittings(ehv_asset_category = "33kV Fittings",
location_risk = "Default",
type_risk = "Default")
}
