% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_lv_ugb.R
\name{safety_cof_lv_ugb}
\alias{safety_cof_lv_ugb}
\title{Safety cost of Failure for LV UGB}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
safety_cof_lv_ugb(lv_asset_category, location_risk, type_risk)
}
\arguments{
\item{lv_asset_category}{String The type of LV asset category}

\item{location_risk}{String Type Financial factor criteria for LV UGB
(cf. section D1.2.1, page 178, CNAIM, 2021).}

\item{type_risk}{String. Asses Financial factor criteria for LV UGB
setting (cf. table 221, page 180, CNAIM, 2021).}
}
\value{
Numeric. Financial consequences of failure for LV UGB
}
\description{
This function calculates safety consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Safety consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
safety_cof_lv_ugb(lv_asset_category = "LV UGB", location_risk = "Default", type_risk = "Default")
}
