% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ohl_cond.R
\name{safety_cof_ohl_cond}
\alias{safety_cof_ohl_cond}
\title{Safety cost of Failure for Overhead Line Conductors}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
safety_cof_ohl_cond(
  ohl_cond_asset_category,
  location_risk,
  type_risk,
  gb_ref_given = NULL
)
}
\arguments{
\item{ohl_cond_asset_category}{String The type of Pole asset category
Options: \code{ohl_cond_asset_category = c("33kV OHL (Tower Line) Conductor",
"66kV OHL (Tower Line) Conductor", "132kV OHL (Tower Line) Conductor")}.}

\item{location_risk}{String Type Financial factor criteria for Overhead Line Conductors
(cf. section D1.2.1, page 178, CNAIM, 2021).
\code{location_risk = c("Low", "Medium", "High")}.
The default setting is
\code{location_risk = "Medium"}.}

\item{type_risk}{String. Asses Financial factor criteria for Overhead Line Conductors
setting (cf. table 221, page 180, CNAIM, 2021).
\code{type_risk = c("Low", "Medium", "High")}.
The default setting is
\code{type_risk = "Medium"}.}

\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Safety consequences of failure for Overhead Line Conductors
}
\description{
This function calculates safety consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Safety consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
safety_cof_ohl_cond(
ohl_cond_asset_category = "33kV OHL (Tower Line) Conductor",
location_risk = "Default",
type_risk = "Default")
}
