% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_relay.R
\name{safety_cof_relay}
\alias{safety_cof_relay}
\title{Safety cost of Failure for Relays}
\usage{
safety_cof_relay(location_risk, type_risk)
}
\arguments{
\item{location_risk}{String Type Financial factor criteria for 50kV fittings
Options: \code{location_risk = c("Low", "Medium", "High")}.
The default setting is
\code{location_risk = "Medium"}.}

\item{type_risk}{String. Asses Financial factor criteria for 50kV fittings
setting
Options: \code{type_risk = c("Low", "Medium", "High")}.
The default setting is
\code{type_risk = "Medium"}.}
}
\value{
Numeric. Financial consequences of failure for relay
}
\description{
This function calculates safety consequences of failure.
Safety consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
