% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_serviceline.R
\name{safety_cof_serviceline}
\alias{safety_cof_serviceline}
\title{Safety cost of Failure for Service Lines}
\usage{
safety_cof_serviceline(gb_ref_given = NULL)
}
\arguments{
\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Financial consequences of failure for service line
}
\description{
This function calculates safety consequences of failure
Outputted in DKK
}
\examples{
safety_cof_serviceline()
}
