% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_submarine_cables_30_60kv.R
\name{safety_cof_submarine_cables_30_60kv}
\alias{safety_cof_submarine_cables_30_60kv}
\title{Safety cost of Failure for 30kV and 60kV Submarine Cables}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
safety_cof_submarine_cables_30_60kv()
}
\value{
Numeric. Safety consequences of failure for Sub cables
}
\description{
This function calculates safety consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Safety consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
\examples{
safety_cof_submarine_cables_30_60kv()
}
