% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_switchgear_30_60kv.R
\name{safety_cof_switchgear_30_60kv}
\alias{safety_cof_switchgear_30_60kv}
\title{Safety cost of Failure for 30kV and 60kV Switchgear}
\usage{
safety_cof_switchgear_30_60kv(
  ehv_asset_category,
  location_risk,
  type_risk,
  gb_ref_given = NULL
)
}
\arguments{
\item{ehv_asset_category}{String The type of 30kV and 60kV switchgear
Options: \code{ehv_asset_category = c("30kV", "60kV")}.}

\item{location_risk}{String Type Financial factor criteria for 30kV and 60kV switchgear
Options: \code{location_risk = c("Low", "Medium", "High")}.
The default setting is
\code{location_risk = "Medium"}.}

\item{type_risk}{String. Asses Financial factor criteria for 30kV and 60kV switchgear
setting.
Options: \code{type_risk = c("Low", "Medium", "High")}.
The default setting is
\code{type_risk = "Medium"}.}

\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Financial consequences of failure for 30kV and 60kV switchgear
}
\description{
This function calculates safety consequences of failure
Safety consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
\examples{
safety_cof_switchgear_30_60kv(ehv_asset_category = "30kV",
location_risk = "Default",
type_risk = "Default")
}
