% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_tower_ohl_support_50kv.R
\name{safety_cof_tower_ohl_support_50kv}
\alias{safety_cof_tower_ohl_support_50kv}
\title{Safety cost of Failure for Tower OHL Support 50 kV}
\usage{
safety_cof_tower_ohl_support_50kv(
  location_risk,
  type_risk,
  gb_ref_given = NULL
)
}
\arguments{
\item{location_risk}{String Type Financial factor criteria for tower
Options: \code{location_risk = c("Low", "Medium", "High")}.
The default setting is
\code{location_risk = "Medium"}.}

\item{type_risk}{String. Asses Financial factor criteria for tower
Options: \code{type_risk = c("Low", "Medium", "High")}.
The default setting is
\code{type_risk = "Medium"}.}

\item{gb_ref_given}{optional parameter to use custom reference values}
}
\value{
Numeric. Safety consequences of failure for tower ohl support 50 kV
}
\description{
This function calculates safety consequences of failure
Safety consequences of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
Outputted in DKK.
}
\examples{
safety_cof_tower_ohl_support_50kv(
location_risk = "Default",
type_risk = "Default")
}
