% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof.R
\name{cof}
\alias{cof}
\title{Consequences of Failure}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/system/files/docs/2017/05/dno_common_network_asset_indices_methodology_v1.1.pdf}
}
\usage{
cof(financial_cof, safety_cof, environmental_cof, network_cof)
}
\arguments{
\item{financial_cof}{Numeric. Financial consequences of failure.}

\item{safety_cof}{Numeric. Safety consequences of failure.}

\item{environmental_cof}{Numeric. Environmental consequences of failure.}

\item{network_cof}{Numeric. Network cost of failure.}
}
\value{
Numeric. Consequences of failure.
}
\description{
This function calculates consequences of failure
(cf.section 7, page 71, CNAIM, 2017).
}
\examples{
# Consequences of failure
# cof(financial_cof, safety_cof, environmental_cof, network_cof)

}
