% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_lv_switchgear_and_other.R
\name{environmental_cof_lv_switchgear_and_other}
\alias{environmental_cof_lv_switchgear_and_other}
\title{Environmental cost of Failure for LV swicthgear and others}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
environmental_cof_lv_switchgear_and_other(lv_asset_category)
}
\arguments{
\item{lv_asset_category}{String The type of LV asset category
Options: \code{lv_asset_category = c("LV Board (WM)",
"LV Board (X-type Network) (WM)", "LV Circuit Breaker",
"LV Pillar (ID)", "LV Pillar (OD at Substation)", "LV Pillar (OD not at a Substation)")}}
}
\description{
This function calculates environmental consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Environmental consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().#' @return Numeric. Financial consequences of failure for LV switchgear
}
\examples{
environmental_cof_lv_switchgear_and_other(lv_asset_category = "LV Board (WM)")
}
