% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ohl_cond.R
\name{environmental_cof_ohl_cond}
\alias{environmental_cof_ohl_cond}
\title{Environmental cost of Failure for Overhead line conductors}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
environmental_cof_ohl_cond(ohl_cond_asset_category)
}
\arguments{
\item{ohl_cond_asset_category}{String The type of Overhead Line Conductors}
}
\description{
This function calculates environmental consequences of failure
(cf. section 7.3, page 75, CNAIM, 2017). Environmental consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().#' @return Numeric. Financial consequences of failure for LV switchgear
}
\examples{
environmental_cof_ohl_cond(ohl_cond_asset_category = "33kV OHL (Tower Line) Conductor")
}
