% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_ehv_fittings.R
\name{financial_cof_ehv_fittings}
\alias{financial_cof_ehv_fittings}
\title{Financial cost of Failure for EHV/132kV fittings}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
financial_cof_ehv_fittings(
  ehv_asset_category,
  type_financial_factor_criteria,
  access_factor_criteria
)
}
\arguments{
\item{ehv_asset_category}{String The type of EHV asset category
Options: \code{ehv_asset_category = c("33kV Fittings", "66kV Fittings",
 "132kV Fittings")}}

\item{type_financial_factor_criteria}{String. Type Financial factor criteria for EHV fittings
\code{type_financial_factor_criteria = c("Suspension", "Tension")}.}

\item{access_factor_criteria}{String. Asses Financial factor criteria for EHV fittings
setting (cf. table 221, page 180, CNAIM, 2021).
\code{access_factor_criteria = c("Type A", "Type B")}.}
}
\value{
Numeric. Financial consequences of failure for EHV fittings
}
\description{
This function calculates financial consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Financial consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
financial_cof_ehv_fittings(ehv_asset_category = "33kV Fittings",
type_financial_factor_criteria = "Tension",
access_factor_criteria = "Type A")
}
