% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_hv_poles.R
\name{pof_future_poles}
\alias{pof_future_poles}
\title{Future Probability of Failure for Poles}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
pof_future_poles(
  pole_asset_category = "20kV Poles",
  sub_division = "Wood",
  placement = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  pole_decay = "default",
  observed_condition_inputs,
  reliability_factor = "Default",
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric array. Future probability of failure
per annum per kilometre for poles.
}
\description{
This function calculates the future
annual probability of failure per kilometer for a poles.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function. For more information about the
probability of failure function see section 6
on page 34 in CNAIM (2021).
}
\examples{
# Future annual probability of failure for HV Poles
}
