% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof_future_transformer_04_10kv.R
\name{pof_future_transformer_04_10kv}
\alias{pof_future_transformer_04_10kv}
\title{Future Probability of Failure for 0.4/10kV Transformers}
\usage{
pof_future_transformer_04_10kv(
  utilisation_pct = "Default",
  placement = "Default",
  altitude_m = "Default",
  distance_from_coast_km = "Default",
  corrosion_category_index = "Default",
  age,
  partial_discharge = "Default",
  temperature_reading = "Default",
  observed_condition = "Default",
  reliability_factor = "Default",
  moisture = "Default",
  acidity = "Default",
  bd_strength = "Default",
  k_value = 0.0077,
  c_value = 1.087,
  normal_expected_life = 55,
  simulation_end_year = 100
)
}
\arguments{
\item{simulation_end_year}{Numeric. The last year of simulating probability
of failure. Default is 100.}
}
\value{
Numeric array. Future probability of failure.
}
\description{
This function calculates the future
annual probability of failure for 0.4/10kV Transformers.
The function is a cubic curve that is based on
the first three terms of the Taylor series for an
exponential function.
}
\examples{
# Future probability of a 0.4/10 kV transformer
future_pof_transformer <-
pof_future_transformer_04_10kv(utilisation_pct = "Default",
placement = "Default",
altitude_m = "Default",
distance_from_coast_km = "Default",
corrosion_category_index = "Default",
age = 20,
partial_discharge = "Default",
temperature_reading = "Default",
observed_condition = "Default",
reliability_factor = "Default",
moisture = "Default",
acidity = "Default",
bd_strength = "Default",
k_value = 0.0077,
c_value = 1.087,
normal_expected_life = 55,
simulation_end_year = 100)
}
