% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_value_of_future_risk.R
\name{present_value_future_risk}
\alias{present_value_future_risk}
\title{Present Value of Future Risk}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 2.1, 2021:
\url{https://www.ofgem.gov.uk/sites/default/files/docs/2021/04/dno_common_network_asset_indices_methodology_v2.1_final_01-04-2021.pdf}
}
\usage{
present_value_future_risk(pof, cof, r = 0.035)
}
\arguments{
\item{pof}{A vector of the probability of failure of the asset over years}

\item{cof}{The consequence of failure of the asset}

\item{r}{discount rate}
}
\description{
This function calculates the present value of future risk.
See section 5.5 on page 32 in CNAIM (2021).
}
\examples{
present_value_future_risk(c(0.1, 0.2, 0.5), 100)
}
