% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_hv_switchgear_distribution.R
\name{safety_cof_hv_switchgear_distribution}
\alias{safety_cof_hv_switchgear_distribution}
\title{Safety cost of Failure for HV Switchgear Distribution}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/system/files/docs/2017/05/dno_common_network_asset_indices_methodology_v1.1.pdf}
}
\usage{
safety_cof_hv_switchgear_distribution(
  hv_asset_category,
  location_risk,
  type_risk
)
}
\arguments{
\item{hv_asset_category}{String The type of LV asset category}

\item{location_risk}{String Type Financial factor criteria for LV switchgear
(cf. section D1.2.1, page 162, CNAIM, 2017).}

\item{type_risk}{String. Asses Financial factor criteria for LV switchgear
setting (cf. table 214, page 164, CNAIM, 2017).}
}
\value{
Numeric. Financial consequences of failure for LV switchgear
}
\description{
This function calculates safety consequences of failure
(cf. section 7.3, page 75, CNAIM, 2017). Safetyr consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
safety_cof_hv_switchgear_distribution(
hv_asset_category = "6.6/11kV CB (GM) Secondary",
location_risk = "Default",
type_risk = "Default")
}
