% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cof_transformers.R
\name{safety_cof_transformers}
\alias{safety_cof_transformers}
\title{Safety cost of Failure for Transformer}
\source{
DNO Common Network Asset Indices Methodology (CNAIM),
Health & Criticality - Version 1.1, 2017:
\url{https://www.ofgem.gov.uk/system/files/docs/2017/05/dno_common_network_asset_indices_methodology_v1.1.pdf}
}
\usage{
safety_cof_transformers(tf_asset_category, location_risk, type_risk)
}
\arguments{
\item{tf_asset_category}{String The type of Transformer}

\item{location_risk}{String Type Financial factor criteria for Transformer
(cf. section D1.2.1, page 178, CNAIM, 2021).}

\item{type_risk}{String. Asses Financial factor criteria for Transformer
setting (cf. table 221, page 180, CNAIM, 2021).}
}
\value{
Numeric. Safety consequences of failure for Transformers
}
\description{
This function calculates safety consequences of failure
(cf. section 7.3, page 79, CNAIM, 2021). Safety consequences
of failure is used in
the derivation of consequences of failure see \code{\link{cof}}().
}
\examples{
safety_cof_transformers(tf_asset_category = "33kV Transformer (GM)",
location_risk = "Default",
type_risk = "Default")
}
